requirements.txt
sentence-transformers
langchain
flask
pypdf
python-dotenv
pinecone[grpc]
langchain-pinecone
langchain_community
langchain_openai
langchain_experimental
-e .
# FastAPI and ASGI server
fastapi
uvicorn[standard]
python-multipart

# Deep Learning
tensorflow
numpy
Pillow
# Development tools (optional)
pytest
httpx